refactor(build-compatible-ext): resolve JPMS split packages
Move classes from dev.langchain4j.cdi.aiservice, .plugin, and .spi
into dev.langchain4j.cdi.core.buildcompatibleextension to eliminate
split package violations between langchain4j-cdi-core and
langchain4j-cdi-build-compatible-ext. Updates META-INF/services
registrations and test imports accordingly.
