+++ langchain4j-cdi-build-compatible-ext/src/main/java/dev/langchain4j/cdi/core/buildcompatibleextension/AIServiceCreator.java
@@ -1,5 +1,6 @@
-package dev.langchain4j.cdi.aiservice;
+package dev.langchain4j.cdi.core.buildcompatibleextension;
 
+import dev.langchain4j.cdi.aiservice.CommonAIServiceCreator;
 import jakarta.enterprise.inject.Instance;
 import jakarta.enterprise.inject.build.compatible.spi.Parameters;
 import jakarta.enterprise.inject.build.compatible.spi.SyntheticBeanCreator;
