+++ langchain4j-cdi-build-compatible-ext/src/main/resources/META-INF/services/jakarta.enterprise.inject.build.compatible.spi.BuildCompatibleExtension
@@ -1,2 +1,2 @@
-dev.langchain4j.cdi.aiservice.Langchain4JAIServiceBuildCompatibleExtension
-dev.langchain4j.cdi.plugin.LangChain4JPluginsBuildCompatibleExtension
+dev.langchain4j.cdi.core.buildcompatibleextension.Langchain4JAIServiceBuildCompatibleExtension
+dev.langchain4j.cdi.core.buildcompatibleextension.LangChain4JPluginsBuildCompatibleExtension
